Group core modal components
Modal views and templates were generic additions to core, but since now
we want to add models as well, it is best to group them together in
a sub-directory.
