+++ src/mmw/js/src/core/modals/templates/confirmModal.html
