+++ src/mmw/js/src/core/modals/templates/inputModal.html
