+++ src/mmw/js/src/core/modals/templates/shareModal.html
