+++ src/mmw/js/src/core/modals/views.js
@@ -1,0 +1,162 @@
+"use strict";
+
+var _ = require('underscore'),
+    Marionette = require('../../../shim/backbone.marionette'),
+    ZeroClipboard = require('zeroclipboard'),
+    models = require('./models'),
+    modalConfirmTmpl = require('./templates/confirmModal.html'),
+    modalInputTmpl = require('./templates/inputModal.html'),
+    modalShareTmpl = require('./templates/shareModal.html'),
+
+    ENTER_KEYCODE = 13,
+    BASIC_MODAL_CLASS = 'modal modal-basic fade';
+
+var ModalBaseView = Marionette.ItemView.extend({
+    className: BASIC_MODAL_CLASS,
+
+    attributes: {
+        'tabindex': '-1'
+    },
+
+    events: {
+        'shown.bs.modal': 'onModalShown',
+        'keyup': 'onKeyUp',
+        'hidden.bs.modal': 'onModalHidden'
+    },
+
+    onRender: function() {
+        this.$el.modal('show');
+    },
+
+    onModalShown: function() {
+        // Not implemented.
+    },
+
+    onKeyUp: function(e) {
+        if (e.keyCode === ENTER_KEYCODE) {
+            this.primaryAction();
+        }
+    },
+
+    primaryAction: function() {
+        // Not implemented.
+    },
+
+    onModalHidden: function() {
+        this.destroy();
+    },
+
+    hide: function() {
+        this.$el.modal('hide');
+    }
+});
+
+var ConfirmView = ModalBaseView.extend({
+    template: modalConfirmTmpl,
+
+    ui: {
+        confirmation: '.confirm'
+    },
+
+    events: _.defaults({
+        'click @ui.confirmation': 'primaryAction'
+    }, ModalBaseView.prototype.events),
+
+    primaryAction: function() {
+        this.triggerMethod('confirmation');
+        this.hide();
+    }
+});
+
+var InputView = ModalBaseView.extend({
+    template: modalInputTmpl,
+
+    ui: {
+        save: '.save',
+        input: 'input',
+        error: '.error'
+    },
+
+    events: _.defaults({
+        'click @ui.save': 'primaryAction'
+    }, ModalBaseView.prototype.events),
+
+    onModalShown: function() {
+        this.ui.input.focus().select();
+    },
+
+    primaryAction: function() {
+        var val = this.ui.input.val().trim();
+        if (val) {
+            this.triggerMethod('update', val);
+            this.hide();
+        } else {
+            this.ui.error.text('Please enter a valid project name');
+        }
+    }
+});
+
+var ShareView = ModalBaseView.extend({
+    template: modalShareTmpl,
+
+    ui: {
+        'signin': '.signin',
+        'copy': '.copy',
+        'input': 'input'
+    },
+
+    events: _.defaults({
+        'click @ui.signin': 'signIn'
+    }, ModalBaseView.prototype.events),
+
+    initialize: function() {
+        this.zc = new ZeroClipboard();
+    },
+
+    // Override to attach ZeroClipboard to ui.copy button
+    onRender: function() {
+        var self = this;
+        this.$el.on('shown.bs.modal', function() {
+            self.zc.clip(self.ui.copy);
+        });
+
+        if (this.model.get('is_private')) {
+            var question = 'This project is currently private. ' +
+                    'You must make it public before it can be shared with others. ' +
+                    'Once public, anyone with the URL can access it.',
+                confirm = new ConfirmView({
+                    model: new models.ConfirmModel({
+                        question: question,
+                        confirmLabel: 'Make Public',
+                        cancelLabel: 'Cancel'
+                    })
+                });
+
+            confirm.render();
+            confirm.on('confirmation', function() {
+                var project = self.options.app.currProject;
+                project.set('is_private', false);
+                project.saveProjectAndScenarios();
+
+                self.$el.modal('show');
+            });
+        } else {
+            this.$el.modal('show');
+        }
+    },
+
+    onModalShown: function() {
+        this.ui.input.focus().select();
+    },
+
+    signIn: function() {
+        this.options.app.getUserOrShowLogin();
+    }
+
+});
+
+module.exports = {
+    ShareView: ShareView,
+    InputView: InputView,
+    ConfirmView: ConfirmView
+};
